Move tests for commands into subfolder
